fullstack-rag/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Jolkkonen, Johannes (External) <johannes.jolkkonen@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"

fastapi = "^0.110.2"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
langchain = "^0.1.16"
qdrant-client = "^1.9.0"
python-decouple = "^3.8"
openai = "^1.23.6"
langchain-community = "^0.0.34"
bs4 = "^0.0.2"
tiktoken = "^0.6.0"
websockets = "^12.0"
modal = "^0.62.197"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
